refactor(arch): split ui/slash.rs by command group
`src/ui/slash.rs` (2537 lines) had one giant `handle_slash`
function whose match statement spanned ~25 slash-command arms.
Splitting by command family so a developer fixing a `/sessions`
bug doesn't scroll past `/worktree`, `/mcp`, `/btw`, etc.

Conversion:
- `src/ui/slash.rs` → `src/ui/slash/mod.rs` (via git mv —
  preserves history; the `crate::ui::slash` path stays valid)
- New `src/ui/slash/cmd_model.rs` (395 lines) —
  /model, /mode, /prompt, /reasoning, /regen-prompts, /toggle
- New `src/ui/slash/cmd_session.rs` (363 lines) —
  /sessions, /tree, /fork, /clone, /undo, /retry, /tasks, /clear
- New `src/ui/slash/cmd_worktree.rs` (175 lines) —
  /worktree, /wt-merge, /wt-exit
- New `src/ui/slash/cmd_misc.rs` (651 lines) —
  /mcp, /btw, /cd, /panel, /quit, /help, /allow, /loop

mod.rs is now 1048 lines (down from 2537, −59%). The dispatcher
is a clean ~75-line match where each arm is a one-liner call
into the topic file.

Key design choice: introduced `pub(super) struct SlashCtx<'a>`
that bundles the mutable references `handle_slash` was juggling
(renderer, session, cli, cfg, context, agent, client, plus
feature-gated loop_state / mcp_manager / semantic_manager /
lsp_manager). Each handler takes `&mut SlashCtx<'_>` so
signatures stay uniform.

Preserved invariants:
- `handle_slash`, `handle_compress`, `undo_last`, `CompressOutcome`,
  `UndoOutcome`, `align_cut_to_user_boundary`, `try_complete`,
  `slash_command_names`, `is_known_slash_command`,
  `builtin_commands`, `format_completion_preview` — all kept in
  `mod.rs` (the callers from `src/ui/mod.rs` are unchanged).
- `/compress` | `/compact` arm kept in mod.rs since it just
  emits the `DEFER_COMPRESS:` sentinel for the outer loop.
- All `DEFER_*` sentinel errors (`DEFER_COMPRESS`, `DEFER_WT_MERGE`,
  `DEFER_WT_EXIT`) preserved verbatim.

Build clean with full features, --no-default-features, and the
git-worktree+loop subset. Full test suite: 1718 pass / 0 fail
/ 0 ignored.
