--- src/ui/slash.rs
@@ -1,2537 +1,0 @@
-use compact_str::CompactString;
-use crossterm::style::Color;
-use smallvec::SmallVec;
-
-use crate::cli::Cli;
-use crate::config::Config;
-use crate::context::ContextFiles;
-#[cfg(feature = "mcp")]
-use crate::extras::mcp::McpClientManager;
-use crate::permission::SecurityMode;
-use crate::permission::ask::AskSender;
-use crate::permission::checker::PermCheck;
-use crate::provider::{AnyAgent, AnyClient};
-use crate::sandbox::Sandbox;
-#[cfg(feature = "semantic")]
-use crate::semantic::SemanticManager;
-use crate::session::{MessageRole, Session};
-use crate::ui::events::{format_time, render_session};
-use crate::ui::input::InputEditor;
-use crate::ui::renderer::Renderer;
-use crate::ui::theme;
-use crate::ui::tree::{self, short_id};
-
-#[inline]
-fn c_agent() -> Color {
-    theme::agent()
-}
-#[inline]
-fn c_result() -> Color {
-    theme::result()
-}
-#[inline]
-fn c_error() -> Color {
-    theme::error()
-}
-
-/// Walk `cut_idx` forward until the message at that index is a
-/// `User` message (or the index reaches `messages.len()`). This
-/// guarantees the kept tail after compress starts with a User
-/// message, which is what every provider expects after a System
-/// summary. If `cut_idx` already points at a User message or is
-/// past the end, no change. If no user message exists in the
-/// tail, return `messages.len()` — caller surfaces the "nothing to
-/// compress" message.
-///
-/// Matches opencode's `splitTurn` discipline
-/// (`session/compaction.ts:161-184`).
-fn align_cut_to_user_boundary(
-    messages: &[crate::session::SessionMessage],
-    cut_idx: usize,
-) -> usize {
-    let mut i = cut_idx;
-    while i < messages.len() && messages[i].role != MessageRole::User {
-        i += 1;
-    }
-    i
-}
-
-/// Outcome of `undo_last`. `removed` is the number of messages popped;
-/// `had_tool_calls` is set when at least one of the popped messages
-/// had tool calls attached — the caller should surface a warning
-/// because tool side effects (file writes, bash, MCP calls) are NOT
-/// reverted by undo.
-#[derive(Debug, Default)]
-pub struct UndoOutcome {
-    pub removed: usize,
-    pub had_tool_calls: bool,
-}
-
-pub fn undo_last(session: &mut Session) -> UndoOutcome {
-    let len = session.messages.len();
-    if len == 0 {
-        return UndoOutcome::default();
-    }
-    let mut outcome = UndoOutcome::default();
-    let pop = |session: &mut Session, outcome: &mut UndoOutcome| {
-        if let Some(last) = session.messages.last()
-            && !last.tool_calls.is_empty()
-        {
-            outcome.had_tool_calls = true;
-        }
-        session.pop_last_message();
-        outcome.removed += 1;
-    };
-    // Route through `pop_last_message` so the tree + message_store
-    // stay in sync — P4c made direct .messages.pop() incorrect for
-    // branched sessions.
-    if session.messages[len - 1].role == MessageRole::Assistant {
-        pop(session, &mut outcome);
-        if session
-            .messages
-            .last()
-            .is_some_and(|m| m.role == MessageRole::User)
-        {
-            pop(session, &mut outcome);
-        }
-        return outcome;
-    }
-    if session.messages[len - 1].role == MessageRole::User {
-        pop(session, &mut outcome);
-    }
-    outcome
-}
-
-/// Result of an attempted compression. `Compacted` means messages
-/// were actually replaced; `NoOp` covers every path that returned
-/// without shrinking the session (already-within-limits, nothing to
-/// cut, summary too large). Callers driving auto-recovery (the
-/// `ContextOverflow` handler) MUST distinguish these — respawning
-/// the run against an unchanged history just re-emits the same
-/// ContextLength error and loops.
-pub enum CompressOutcome {
-    Compacted,
-    NoOp { reason: &'static str },
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn handle_compress(
-    instructions: Option<&str>,
-    agent: &mut AnyAgent,
-    client: &AnyClient,
-    renderer: &mut Renderer,
-    session: &mut Session,
-    cli: &Cli,
-    cfg: &Config,
-    context: &mut ContextFiles,
-    permission: &Option<PermCheck>,
-    ask_tx: &Option<AskSender>,
-    // Audit followup (companion to C8 LSP fix): question_tx +
-    // plan_tx were previously passed as None when this function
-    // rebuilt the agent post-compact. Tools that depend on either
-    // (the `question` tool, plan-mode switch hooks) silently
-    // broke after every auto-compact + manual /compress. Thread
-    // the channels through.
-    question_tx: &Option<crate::agent::tools::question::QuestionSender>,
-    plan_tx: &Option<crate::agent::tools::plan::PlanSwitchSender>,
-    bg_store: &Option<crate::agent::tools::background::BackgroundStore>,
-    sandbox: &Sandbox,
-    #[cfg(feature = "mcp")] mcp_manager: Option<&McpClientManager>,
-    #[cfg(feature = "semantic")] semantic_manager: Option<&SemanticManager>,
-    #[cfg(feature = "lsp")] lsp_manager: Option<&std::sync::Arc<crate::lsp::manager::LspManager>>,
-) -> anyhow::Result<CompressOutcome> {
-    renderer.write_line("compressing...", c_agent())?;
-    renderer.write_line("", Color::White)?;
-
-    let reserve = cfg.resolve_reserve_tokens();
-    let keep_recent = cfg.resolve_keep_recent_tokens();
-    let max_tokens = session.context_window.saturating_sub(reserve);
-
-    if session.total_estimated_tokens <= max_tokens {
-        renderer.write_line("context within limits, no compression needed", c_agent())?;
-        return Ok(CompressOutcome::NoOp {
-            reason: "context within limits",
-        });
-    }
-
-    let mut accumulated = 0u64;
-    let mut cut_idx = session.messages.len();
-    for (i, msg) in session.messages.iter().enumerate().rev() {
-        if accumulated >= keep_recent {
-            cut_idx = i + 1;
-            break;
-        }
-        accumulated = accumulated.saturating_add(msg.estimated_tokens);
-    }
-
-    // F3: nudge `cut_idx` forward until the first kept message is a
-    // User message (or we hit the end). Without this, the kept tail
-    // could start with an Assistant message — Anthropic + OpenAI
-    // expect alternating user/assistant role order; an assistant
-    // following a system summary breaks the role sequence and gets
-    // rejected with a 400. opencode handles this via `splitTurn`
-    // (`compaction.ts:161-184`).
-    cut_idx = align_cut_to_user_boundary(&session.messages, cut_idx);
-
-    if cut_idx == 0 {
-        renderer.write_line("nothing to compress (entire context is recent)", c_agent())?;
-        return Ok(CompressOutcome::NoOp {
-            reason: "entire context is within keep_recent_tokens — lower it to compress further",
-        });
-    }
-
-    let messages_to_summarize = &session.messages[..cut_idx];
-    let previous_summary = session.compactions.last().map(|c| c.summary.as_str());
-
-    let summary = client
-        .compress_messages(
-            &session.model,
-            messages_to_summarize,
-            previous_summary,
-            instructions,
-        )
-        .await?;
-
-    let tokens_before: u64 = messages_to_summarize
-        .iter()
-        .map(|m| m.estimated_tokens)
-        .sum();
-
-    // F13: estimate the summary's own token cost so we can
-    // report TRUE net savings instead of just "tokens replaced".
-    // A pathological summary longer than the messages it
-    // replaces means we just paid more tokens for less context.
-    // We still proceed with the compress (the new prefix is the
-    // SHAPE the LLM expects), but we want to surface the
-    // misfire so the user can adjust `keep_recent_tokens` or
-    // their custom compress prompt. opencode validates the
-    // summary fits the budget BEFORE issuing the LLM call
-    // (`compaction.ts:136-294`); dirge validates AFTER because
-    // we don't know the summary's size until the LLM returns.
-    let summary_tokens_est = crate::session::Session::estimate_tokens(&summary);
-    let net_saved: i64 = tokens_before as i64 - summary_tokens_est as i64;
-
-    // Audit M9: previously the summary was installed via
-    // `compress_reporting` BEFORE the net-saved check, so an
-    // oversized summary still landed in the session — we only
-    // told the user *afterwards*. Refuse to install when the
-    // summary would cost more than the messages it replaces; the
-    // user can adjust `keep_recent_tokens` / their compress prompt
-    // and re-issue. Skipping the install also avoids polluting the
-    // session-tree with a node we'd want to revert.
-    if net_saved < 0 {
-        renderer.write_line(
-            &format!(
-                "compress aborted — summary ({}t) is LARGER than the {} messages it would replace ({}t); net cost +{}t. Compression rejected. Consider lowering keep_recent_tokens or refining compress instructions, then re-run /compress.",
-                summary_tokens_est,
-                cut_idx,
-                tokens_before,
-                -net_saved,
-            ),
-            c_error(),
-        )?;
-        return Ok(CompressOutcome::NoOp {
-            reason: "summary would be larger than the messages it replaces",
-        });
-    }
-
-    // `compress_reporting` returns the count of non-active-path
-    // tree nodes (sibling branches) pruned. We notify the user
-    // about that loss explicitly — without the notification a
-    // branched session could silently lose forks during auto-
-    // compaction. opencode (`session/compaction.ts:386-396`) drops
-    // siblings silently; dirge prefers the explicit notification.
-    let pruned_branches = session.compress_reporting(summary, cut_idx, tokens_before);
-
-    let model = client.completion_model(session.model.to_string());
-    *agent = crate::provider::build_agent(
-        model,
-        cli,
-        cfg,
-        context,
-        permission.clone(),
-        ask_tx.clone(),
-        question_tx.clone(),
-        plan_tx.clone(),
-        bg_store.clone(),
-        #[cfg(feature = "lsp")]
-        lsp_manager.cloned(),
-        sandbox.clone(),
-        #[cfg(feature = "mcp")]
-        mcp_manager,
-        #[cfg(feature = "semantic")]
-        semantic_manager,
-    )
-    .await;
-    renderer.write_line("prompt cleared (back to default behavior)", c_agent())?;
-
-    render_session(renderer, session, cli, cfg, context)?;
-    if pruned_branches > 0 {
-        // Tell the user the branched topology shrunk. Without this,
-        // they'd notice missing forks in `/tree` without any
-        // explanation.
-        renderer.write_line(
-            &format!(
-                "discarded {} forked branch node{} that were rooted in the compressed region",
-                pruned_branches,
-                if pruned_branches == 1 { "" } else { "s" },
-            ),
-            c_error(),
-        )?;
-    }
-    // Net-saved is guaranteed non-negative here: the early-return
-    // above (audit M9) aborts the compress when the summary would
-    // cost more than the messages it replaced.
-    {
-        renderer.write_line(
-            &format!(
-                "compressed {} messages (saved ~{} tokens; summary uses {}t)",
-                cut_idx, net_saved, summary_tokens_est,
-            ),
-            c_agent(),
-        )?;
-    }
-
-    Ok(CompressOutcome::Compacted)
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn handle_slash(
-    text: &str,
-    agent: &mut AnyAgent,
-    client: &AnyClient,
-    renderer: &mut Renderer,
-    session: &mut Session,
-    cli: &Cli,
-    cfg: &Config,
-    context: &mut ContextFiles,
-    show_reasoning: &mut bool,
-    is_running: &mut bool,
-    input: &mut InputEditor,
-    permission: &Option<PermCheck>,
-    ask_tx: &Option<AskSender>,
-    // Audit followup: same threading as ask_tx — every build_agent
-    // rebuild inside handle_slash previously passed None for
-    // question_tx + plan_tx, silently killing the `question` tool
-    // and plan-switch hooks after any rebuild-triggering slash
-    // command. Companion to the C8 LSP fix.
-    question_tx: &Option<crate::agent::tools::question::QuestionSender>,
-    plan_tx: &Option<crate::agent::tools::plan::PlanSwitchSender>,
-    todo_tools_enabled: &mut bool,
-    bg_store: &Option<crate::agent::tools::background::BackgroundStore>,
-    sandbox: &Sandbox,
-    #[cfg(feature = "loop")] loop_state: &mut Option<crate::extras::r#loop::LoopState>,
-    #[cfg(feature = "mcp")] mcp_manager: Option<&McpClientManager>,
-    #[cfg(feature = "semantic")] semantic_manager: Option<&SemanticManager>,
-    // C8 (audit fix): every prior agent-rebuild path (/model,
-    // /prompt, /mode, /cd, /worktree, /wt-exit, /regen-prompts,
-    // /loop start/stop, /toggle) passed None for lsp_manager into
-    // build_agent. The user lost LSP silently after the first such
-    // command. Thread the actual manager through.
-    #[cfg(feature = "lsp")] lsp_manager: Option<&std::sync::Arc<crate::lsp::manager::LspManager>>,
-) -> anyhow::Result<()> {
-    let parts: SmallVec<[&str; 3]> = text.trim().splitn(3, ' ').collect();
-    match parts[0] {
-        "/model" => {
-            if parts.len() < 2 {
-                renderer.write_line(&format!("current model: {}", session.model), c_agent())?;
-            } else {
-                let new_model = CompactString::new(parts[1].trim());
-                let model = client.completion_model(new_model.to_string());
-                *agent = crate::provider::build_agent(
-                    model,
-                    cli,
-                    cfg,
-                    context,
-                    permission.clone(),
-                    ask_tx.clone(),
-                    question_tx.clone(),
-                    plan_tx.clone(),
-                    bg_store.clone(),
-                    #[cfg(feature = "lsp")]
-                    lsp_manager.cloned(),
-                    sandbox.clone(),
-                    #[cfg(feature = "mcp")]
-                    mcp_manager,
-                    #[cfg(feature = "semantic")]
-                    semantic_manager,
-                )
-                .await;
-                session.model = new_model.clone();
-                session.provider = cli.resolve_provider(cfg);
-                // Re-resolve the context window for the new model.
-                // Without this, switching from a 128k model to a 1M
-                // model would leave the status indicator capped at
-                // 128k. config explicit value always wins; otherwise
-                // the per-model table is consulted.
-                let new_ctx = cfg.resolve_context_window(new_model.as_str());
-                let old_ctx = session.context_window;
-                if new_ctx != old_ctx {
-                    session.context_window = new_ctx;
-                }
-                renderer.write_line(&format!("switched to model: {}", new_model), c_agent())?;
-                // Review #3: if the new model has a smaller window
-                // and the current session already exceeds it
-                // (minus reserve), the next prompt would error out
-                // with ContextLength mid-stream. Surface a warning
-                // + recommend `/compress` so the user can shrink
-                // proactively. The auto-compact recovery path also
-                // catches this on the next prompt; the warning
-                // here lets the user act without burning a turn.
-                let reserve = cfg.resolve_reserve_tokens();
-                let budget = new_ctx.saturating_sub(reserve);
-                if new_ctx < old_ctx && session.total_estimated_tokens > budget {
-                    renderer.write_line(
-                        &format!(
-                            "warning: session uses ~{}k tokens but new model's context budget is ~{}k. Run /compress before the next prompt or the next turn may overflow.",
-                            session.total_estimated_tokens / 1_000,
-                            budget / 1_000,
-                        ),
-                        c_error(),
-                    )?;
-                }
-            }
-        }
-        "/sessions" => {
-            if parts.len() < 2 {
-                let sessions = crate::session::storage::find_recent_sessions(20)?;
-                if sessions.is_empty() {
-                    renderer.write_line("no saved sessions", c_agent())?;
-                } else {
-                    renderer
-                        .write_line(&format!("recent sessions ({}):", sessions.len()), c_agent())?;
-                    for s in &sessions {
-                        let last = s
-                            .messages
-                            .last()
-                            .map(|m| {
-                                format!("...{}", &m.content.chars().take(30).collect::<String>())
-                            })
-                            .unwrap_or_default();
-                        let time = format_time(&s.updated_at);
-                        renderer.write_line(
-                            &format!(
-                                "  {}  {}  {}msgs  {}  {}",
-                                &s.id[..8],
-                                time,
-                                s.messages.len(),
-                                s.model,
-                                last
-                            ),
-                            c_result(),
-                        )?;
-                    }
-                }
-            } else if parts[1] == "delete" && parts.len() >= 3 {
-                let prefix = parts[2].trim();
-                let sessions = crate::session::storage::find_sessions_by_prefix(prefix)?;
-                if sessions.is_empty() {
-                    renderer.write_line(&format!("no session matching '{}'", prefix), c_agent())?;
-                } else if sessions.len() == 1 {
-                    if let Some(s) = sessions.into_iter().next() {
-                        let id = s.id.clone();
-                        let preview = s
-                            .messages
-                            .last()
-                            .map(|m| {
-                                format!("...{}", &m.content.chars().take(40).collect::<String>())
-                            })
-                            .unwrap_or_default();
-                        if let Err(e) = crate::session::storage::delete_session(&id) {
-                            renderer.write_line(&format!("failed to delete: {}", e), c_error())?;
-                        } else {
-                            renderer.write_line(
-                                &format!("deleted session {} {}", &id[..8], preview),
-                                c_agent(),
-                            )?;
-                        }
-                    }
-                } else {
-                    renderer.write_line(
-                        &format!("multiple sessions match '{}', be more specific", prefix),
-                        c_agent(),
-                    )?;
-                    for s in &sessions {
-                        let last = s
-                            .messages
-                            .last()
-                            .map(|m| {
-                                format!("...{}", &m.content.chars().take(30).collect::<String>())
-                            })
-                            .unwrap_or_default();
-                        let time = format_time(&s.updated_at);
-                        renderer.write_line(
-                            &format!(
-                                "  {}  {}  {}msgs  {}  {}",
-                                &s.id[..8],
-                                time,
-                                s.messages.len(),
-                                s.model,
-                                last
-                            ),
-                            c_result(),
-                        )?;
-                    }
-                }
-            } else {
-                let prefix = parts[1].trim();
-                let sessions = crate::session::storage::find_sessions_by_prefix(prefix)?;
-                if sessions.is_empty() {
-                    renderer.write_line(&format!("no session matching '{}'", prefix), c_agent())?;
-                } else if sessions.len() == 1 {
-                    if let Some(s) = sessions.into_iter().next() {
-                        let msg_count = s.messages.len();
-                        // Cancel any in-flight background subagent
-                        // tasks belonging to the previous session
-                        // before swapping it out. Otherwise they
-                        // would keep consuming API budget against a
-                        // session their parent agent no longer sees.
-                        if let Some(store) = bg_store.as_ref() {
-                            store.cancel_all();
-                        }
-                        *session = s;
-                        // Restore the prompt that was active when the
-                        // session was saved. Without this, `/sessions
-                        // <id>` would always rebuild the agent with the
-                        // default prompt, even if the saved session
-                        // recorded `current_prompt_name = Some("plan")`.
-                        // Sets both context fields + rebuilds the agent
-                        // so the new prompt actually takes effect.
-                        let restored = session.current_prompt_name.clone();
-                        if let Some(name) = restored.as_deref()
-                            && let Some(p) = context.prompts.get(name)
-                        {
-                            context.current_prompt = Some(p.body.clone());
-                            context.current_prompt_name = Some(name.to_string());
-                            context.current_prompt_deny_tools = p.deny_tools.clone();
-                            crate::permission::apply_prompt_deny(
-                                permission,
-                                &context.current_prompt_deny_tools,
-                            );
-                            let model = client.completion_model(session.model.to_string());
-                            *agent = crate::provider::build_agent(
-                                model,
-                                cli,
-                                cfg,
-                                context,
-                                permission.clone(),
-                                ask_tx.clone(),
-                                question_tx.clone(),
-                                plan_tx.clone(),
-                                bg_store.clone(),
-                                #[cfg(feature = "lsp")]
-                                lsp_manager.cloned(),
-                                sandbox.clone(),
-                                #[cfg(feature = "mcp")]
-                                mcp_manager,
-                                #[cfg(feature = "semantic")]
-                                semantic_manager,
-                            )
-                            .await;
-                        }
-                        render_session(renderer, session, cli, cfg, context)?;
-                        let prompt_note = restored
-                            .map(|n| format!("; prompt: {}", n))
-                            .unwrap_or_default();
-                        renderer.write_line(
-                            &format!("loaded session ({} msgs{})", msg_count, prompt_note),
-                            c_agent(),
-                        )?;
-                    }
-                } else {
-                    renderer
-                        .write_line(&format!("multiple sessions match '{}':", prefix), c_agent())?;
-                    for s in &sessions {
-                        let last = s
-                            .messages
-                            .last()
-                            .map(|m| {
-                                format!("...{}", &m.content.chars().take(30).collect::<String>())
-                            })
-                            .unwrap_or_default();
-                        let time = format_time(&s.updated_at);
-                        renderer.write_line(
-                            &format!(
-                                "  {}  {}  {}msgs  {}  {}",
-                                &s.id[..8],
-                                time,
-                                s.messages.len(),
-                                s.model,
-                                last
-                            ),
-                            c_result(),
-                        )?;
-                    }
-                }
-            }
-        }
-        "/reasoning" => {
-            *show_reasoning = !*show_reasoning;
-            renderer.write_line(
-                &format!(
-                    "reasoning visibility: {}",
-                    if *show_reasoning { "on" } else { "off" }
-                ),
-                c_agent(),
-            )?;
-        }
-        "/mode" => {
-            let current_mode = permission
-                .as_ref()
-                .map(|p| p.lock().unwrap_or_else(|e| e.into_inner()).mode())
-                .unwrap_or(SecurityMode::Standard);
-
-            if parts.len() < 2 {
-                renderer.write_line("security mode:", c_agent())?;
-                renderer.write_line(&format!("  current: {}", current_mode), c_result())?;
-                renderer.write_line("", c_agent())?;
-                renderer.write_line(
-                    "  /mode standard      use configured permission rules",
-                    c_result(),
-                )?;
-                renderer
-                    .write_line("  /mode restrictive   default all tools to ask", c_result())?;
-                renderer.write_line(
-                    "  /mode accept        auto-accept within working directory",
-                    c_result(),
-                )?;
-                renderer.write_line(
-                    "  /mode yolo          auto-accept ALL operations",
-                    c_result(),
-                )?;
-                renderer.write_line("", c_agent())?;
-            } else {
-                match parts[1] {
-                    "standard" => {
-                        if let Some(p) = permission {
-                            p.lock()
-                                .unwrap_or_else(|e| e.into_inner())
-                                .set_mode(SecurityMode::Standard);
-                            renderer.write_line("security mode: standard", c_agent())?;
-                        } else {
-                            renderer.write_line("permission system not active", c_error())?;
-                        }
-                    }
-                    "restrictive" => {
-                        if let Some(p) = permission {
-                            p.lock()
-                                .unwrap_or_else(|e| e.into_inner())
-                                .set_mode(SecurityMode::Restrictive);
-                            renderer.write_line("security mode: restrictive", c_agent())?;
-                        } else {
-                            renderer.write_line("permission system not active", c_error())?;
-                        }
-                    }
-                    "accept" => {
-                        if let Some(p) = permission {
-                            p.lock()
-                                .unwrap_or_else(|e| e.into_inner())
-                                .set_mode(SecurityMode::Accept);
-                            renderer.write_line(
-                                "security mode: accept (auto-allow within CWD)",
-                                c_agent(),
-                            )?;
-                        } else {
-                            renderer.write_line("permission system not active", c_error())?;
-                        }
-                    }
-                    "yolo" => {
-                        if let Some(p) = permission {
-                            p.lock()
-                                .unwrap_or_else(|e| e.into_inner())
-                                .set_mode(SecurityMode::Yolo);
-                            renderer.write_line(
-                                "security mode: YOLO (all operations allowed)",
-                                c_agent(),
-                            )?;
-                        } else {
-                            renderer.write_line("permission system not active", c_error())?;
-                        }
-                    }
-                    _ => {
-                        renderer.write_line(&format!("unknown mode: {}", parts[1]), c_error())?;
-                    }
-                }
-            }
-        }
-        #[cfg(feature = "mcp")]
-        "/mcp" => {
-            let Some(mgr) = mcp_manager else {
-                renderer.write_line("no MCP servers configured", c_agent())?;
-                return Ok(());
-            };
-            let connections = mgr.connections_snapshot();
-            if connections.is_empty() {
-                renderer.write_line("no MCP servers connected", c_agent())?;
-            } else if parts.len() == 1 {
-                renderer.write_line("MCP servers:", c_agent())?;
-                for (server_name, conn) in &connections {
-                    match crate::extras::mcp::client::list_tools(conn).await {
-                        Ok(tools) => {
-                            renderer.write_line(
-                                &format!("  {} ({} tools)", server_name, tools.len()),
-                                c_result(),
-                            )?;
-                        }
-                        Err(e) => {
-                            renderer.write_line(
-                                &format!("  {} (error: {})", server_name, e),
-                                c_error(),
-                            )?;
-                        }
-                    }
-                }
-            } else {
-                let name = parts[1].trim();
-                if let Some(conn) = connections.iter().find(|(n, _)| n == name).map(|(_, c)| c) {
-                    match crate::extras::mcp::client::list_tools(conn).await {
-                        Ok(tools) => {
-                            if tools.is_empty() {
-                                renderer.write_line(
-                                    &format!("server '{}' has no tools", name),
-                                    c_agent(),
-                                )?;
-                            } else {
-                                renderer.write_line(&format!("tools on '{}':", name), c_agent())?;
-                                for tool in &tools {
-                                    let desc = tool.description.as_deref().unwrap_or("");
-                                    renderer.write_line(
-                                        &format!("  {}  {}", tool.name, desc),
-                                        c_result(),
-                                    )?;
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            renderer.write_line(
-                                &format!("error listing tools on '{}': {}", name, e),
-                                c_error(),
-                            )?;
-                        }
-                    }
-                } else {
-                    renderer.write_line(&format!("unknown MCP server: '{}'", name), c_error())?;
-                }
-            }
-        }
-        "/toggle" => {
-            if parts.len() < 2 {
-                renderer.write_line("usage: /toggle <feature> [on|off]", c_agent())?;
-                renderer.write_line("features:", c_agent())?;
-                renderer.write_line(
-                    &format!("  todo  {}", if *todo_tools_enabled { "on" } else { "off" }),
-                    c_result(),
-                )?;
-            } else {
-                let new_state = match parts.get(2).copied() {
-                    Some("on") => true,
-                    Some("off") => false,
-                    Some(other) => {
-                        renderer.write_line(
-                            &format!("invalid: '{}', use on or off", other),
-                            c_error(),
-                        )?;
-                        return Ok(());
-                    }
-                    None => !*todo_tools_enabled,
-                };
-                if new_state == *todo_tools_enabled {
-                    renderer.write_line(
-                        &format!(
-                            "todo tools already {}",
-                            if new_state { "on" } else { "off" }
-                        ),
-                        c_agent(),
-                    )?;
-                } else {
-                    *todo_tools_enabled = new_state;
-                    let model = client.completion_model(session.model.to_string());
-                    *agent = crate::provider::build_agent(
-                        model,
-                        cli,
-                        cfg,
-                        context,
-                        permission.clone(),
-                        ask_tx.clone(),
-                        question_tx.clone(),
-                        plan_tx.clone(),
-                        bg_store.clone(),
-                        #[cfg(feature = "lsp")]
-                        lsp_manager.cloned(),
-                        sandbox.clone(),
-                        #[cfg(feature = "mcp")]
-                        mcp_manager,
-                        #[cfg(feature = "semantic")]
-                        semantic_manager,
-                    )
-                    .await;
-                    renderer.write_line(
-                        &format!(
-                            "todo tools: {}",
-                            if *todo_tools_enabled { "on" } else { "off" }
-                        ),
-                        c_agent(),
-                    )?;
-                }
-            }
-        }
-        "/compress" | "/compact" => {
-            let instructions = if parts.len() > 1 {
-                Some(parts[1..].join(" "))
-            } else {
-                None
-            };
-            let instr_str = instructions.clone().unwrap_or_default();
-            return Err(anyhow::anyhow!("DEFER_COMPRESS:{}", instr_str));
-        }
-        "/loop" => {
-            #[cfg(feature = "loop")]
-            {
-                if parts.len() < 2 || (parts.len() >= 2 && parts[1] == "status") {
-                    if let Some(ls) = loop_state {
-                        let status = if ls.active { "active" } else { "stopped" };
-                        renderer.write_line(
-                            &format!(
-                                "loop {}: {} ({})",
-                                status,
-                                ls.iteration_label(),
-                                ls.plan_file.display()
-                            ),
-                            c_agent(),
-                        )?;
-                    } else {
-                        renderer.write_line("no active loop", c_agent())?;
-                        renderer.write_line("usage: /loop <prompt>  |  /loop stop", c_result())?;
-                    }
-                } else if parts[1] == "stop" {
-                    if let Some(ls) = loop_state {
-                        ls.active = false;
-                        renderer.write_line("loop stopped", c_agent())?;
-                    } else {
-                        renderer.write_line("no active loop", c_agent())?;
-                    }
-                } else {
-                    // B3-4 (audit fix): parse `--max N` flag + apply
-                    // a default cap of 20 iterations so a runaway
-                    // /loop can't burn API spend until the user
-                    // notices and types /loop stop. `--max 0`
-                    // explicitly opts into unbounded (matches the
-                    // prior default behaviour).
-                    let after = text.trim().strip_prefix("/loop").unwrap_or("").trim_start();
-                    let tokens: Vec<&str> = after.split_whitespace().collect();
-                    let mut max_iterations: Option<u32> = Some(20); // default cap
-                    let mut prompt_tokens: Vec<&str> = Vec::new();
-                    let mut i = 0;
-                    while i < tokens.len() {
-                        if tokens[i] == "--max" && i + 1 < tokens.len() {
-                            match tokens[i + 1].parse::<u32>() {
-                                Ok(0) => max_iterations = None,
-                                Ok(n) => max_iterations = Some(n),
-                                Err(_) => {
-                                    renderer.write_line(
-                                        &format!(
-                                            "invalid --max value: {} (use a positive integer, or 0 for unbounded)",
-                                            tokens[i + 1]
-                                        ),
-                                        c_error(),
-                                    )?;
-                                    return Ok(());
-                                }
-                            }
-                            i += 2;
-                        } else {
-                            prompt_tokens.push(tokens[i]);
-                            i += 1;
-                        }
-                    }
-                    let prompt = prompt_tokens.join(" ");
-                    if prompt.is_empty() {
-                        renderer.write_line(
-                            "usage: /loop [--max N] <prompt>  (default cap: 20 iterations; --max 0 = unbounded)",
-                            c_error(),
-                        )?;
-                        return Ok(());
-                    }
-                    let plan_file = std::path::PathBuf::from("LOOP_PLAN.md");
-                    let ls = crate::extras::r#loop::LoopState::new(
-                        prompt,
-                        plan_file,
-                        max_iterations,
-                        None,
-                    );
-                    *loop_state = Some(ls);
-                    let cap_msg = match max_iterations {
-                        Some(n) => format!("loop started (max {n} iterations) — iteration 1 will run after this message"),
-                        None => "loop started (unbounded — use /loop stop to cancel) — iteration 1 will run after this message".to_string(),
-                    };
-                    renderer.write_line(&cap_msg, c_agent())?;
-                }
-            }
-            #[cfg(not(feature = "loop"))]
-            {
-                renderer.write_line(
-                    "/loop requires the 'loop' feature: cargo build --features loop",
-                    c_error(),
-                )?;
-            }
-        }
-        "/prompt" => {
-            let mut sorted: Vec<&String> = context.prompts.keys().collect();
-            sorted.sort();
-            if parts.len() < 2 {
-                if sorted.is_empty() {
-                    renderer.write_line("no prompts available", c_agent())?;
-                } else {
-                    let current = context.current_prompt_name.as_deref().unwrap_or("(none)");
-                    renderer.write_line(
-                        &format!("available prompts (current: {}):", current),
-                        c_agent(),
-                    )?;
-                    // #3 fix: surface each prompt's `description`
-                    // frontmatter (if present). The field was being
-                    // parsed but never displayed — a user writing a
-                    // description expected it to show up here.
-                    // Computes a single-pass max-name width for
-                    // alignment when descriptions are present.
-                    let max_name = sorted.iter().map(|n| n.len()).max().unwrap_or(0);
-                    for name in &sorted {
-                        let desc = context
-                            .prompts
-                            .get(*name)
-                            .and_then(|p| p.description.as_deref());
-                        match desc {
-                            Some(d) => renderer.write_line(
-                                &format!("  {:<width$}  {}", name, d, width = max_name),
-                                c_result(),
-                            )?,
-                            None => renderer.write_line(&format!("  {}", name), c_result())?,
-                        }
-                    }
-                    renderer.write_line("", c_agent())?;
-                    renderer.write_line("usage: /prompt <name>  |  /prompt default", c_result())?;
-                }
-            } else if parts[1] == "default" && !context.prompts.contains_key("default") {
-                // Backstop: if there's no `default` prompt file
-                // present (e.g. the user deleted it), preserve the
-                // legacy "clear back to no active prompt" behavior
-                // by accepting `/prompt default` as a clear keyword.
-                // When a `default` prompt IS registered (which is
-                // the README-documented case — `prompts/default.md`
-                // ships in dirge), this branch is skipped and the
-                // normal "activate by name" path below picks it up,
-                // matching user expectation.
-                if context.current_prompt.is_none() {
-                    renderer.write_line("no active prompt to clear", c_agent())?;
-                } else {
-                    context.current_prompt = None;
-                    context.current_prompt_name = None;
-                    context.current_prompt_deny_tools.clear();
-                    crate::permission::apply_prompt_deny(
-                        permission,
-                        &context.current_prompt_deny_tools,
-                    );
-                    // Mirror into session so `-c` / `/sessions <id>`
-                    // resumes the same prompt state next time.
-                    session.current_prompt_name = None;
-                    let model = client.completion_model(session.model.to_string());
-                    *agent = crate::provider::build_agent(
-                        model,
-                        cli,
-                        cfg,
-                        context,
-                        permission.clone(),
-                        ask_tx.clone(),
-                        question_tx.clone(),
-                        plan_tx.clone(),
-                        bg_store.clone(),
-                        #[cfg(feature = "lsp")]
-                        lsp_manager.cloned(),
-                        sandbox.clone(),
-                        #[cfg(feature = "mcp")]
-                        mcp_manager,
-                        #[cfg(feature = "semantic")]
-                        semantic_manager,
-                    )
-                    .await;
-                }
-            } else {
-                let name = parts[1].trim();
-                if let Some(p) = context.prompts.get(name) {
-                    context.current_prompt = Some(p.body.clone());
-                    context.current_prompt_name = Some(name.to_string());
-                    context.current_prompt_deny_tools = p.deny_tools.clone();
-                    crate::permission::apply_prompt_deny(
-                        permission,
-                        &context.current_prompt_deny_tools,
-                    );
-                    // Mirror into the session so resuming restores the
-                    // same active prompt instead of defaulting to "code".
-                    session.current_prompt_name = Some(name.to_string());
-                    let model = client.completion_model(session.model.to_string());
-                    *agent = crate::provider::build_agent(
-                        model,
-                        cli,
-                        cfg,
-                        context,
-                        permission.clone(),
-                        ask_tx.clone(),
-                        question_tx.clone(),
-                        plan_tx.clone(),
-                        bg_store.clone(),
-                        #[cfg(feature = "lsp")]
-                        lsp_manager.cloned(),
-                        sandbox.clone(),
-                        #[cfg(feature = "mcp")]
-                        mcp_manager,
-                        #[cfg(feature = "semantic")]
-                        semantic_manager,
-                    )
-                    .await;
-                    renderer.write_line(&format!("active prompt: {}", name), c_agent())?;
-                } else {
-                    renderer.write_line(&format!("unknown prompt: '{}'", name), c_error())?;
-                    if !sorted.is_empty() {
-                        renderer.write_line("available prompts:", c_agent())?;
-                        for p in &sorted {
-                            renderer.write_line(&format!("  {}", p), c_result())?;
-                        }
-                    }
-                }
-            }
-        }
-        #[cfg(feature = "git-worktree")]
-        "/worktree" => {
-            if parts.len() < 2 {
-                renderer.write_line("usage: /worktree <name>", c_error())?;
-                return Ok(());
-            }
-            let name = parts[1].trim();
-            // EXT-8: fail fast on the obviously-broken / hostile
-            // shapes BEFORE shelling out to git. `git_worktree::create`
-            // performs the canonical check (and rejects more shapes —
-            // reserved refs, `..`, control bytes, etc.); this inline
-            // probe just gives the user an immediate error for the
-            // common typos without paying for a subprocess.
-            let invalid = name.is_empty()
-                || name.contains(' ')
-                || name.contains('/')
-                || name.starts_with('-')
-                || name.contains("..")
-                || name == "HEAD"
-                || name == "@"
-                || name.chars().any(|c| {
-                    c == '\0'
-                        || c == '~'
-                        || c == ':'
-                        || c == '^'
-                        || c == '?'
-                        || c == '*'
-                        || c == '['
-                        || (c.is_control() && c != '\t')
-                });
-            if invalid {
-                renderer.write_line(
-                    "invalid name: use a single word without spaces, slashes, leading '-', '..', or git ref metacharacters (~ : ^ ? * [) — and not 'HEAD' or '@'",
-                    c_error(),
-                )?;
-                return Ok(());
-            }
-
-            match crate::extras::git_worktree::create(name) {
-                Ok((path, _info)) => {
-                    std::env::set_current_dir(&path)
-                        .map_err(|e| anyhow::anyhow!("failed to change directory: {}", e))?;
-                    session.working_dir = compact_str::CompactString::new(path.to_string_lossy());
-                    context.reload();
-                    let model = client.completion_model(session.model.to_string());
-                    *agent = crate::provider::build_agent(
-                        model,
-                        cli,
-                        cfg,
-                        context,
-                        permission.clone(),
-                        ask_tx.clone(),
-                        question_tx.clone(),
-                        plan_tx.clone(),
-                        bg_store.clone(),
-                        #[cfg(feature = "lsp")]
-                        lsp_manager.cloned(),
-                        sandbox.clone(),
-                        #[cfg(feature = "mcp")]
-                        mcp_manager,
-                        #[cfg(feature = "semantic")]
-                        semantic_manager,
-                    )
-                    .await;
-                    render_session(renderer, session, cli, cfg, context)?;
-                    renderer.write_line(
-                        &format!("worktree created: branch '{}' at {}", name, path.display()),
-                        c_agent(),
-                    )?;
-                }
-                Err(e) => {
-                    renderer.write_line(&format!("failed: {}", e), c_error())?;
-                }
-            }
-        }
-        #[cfg(feature = "git-worktree")]
-        "/wt-merge" => {
-            let info = match crate::extras::git_worktree::detect() {
-                Some(i) => i,
-                None => {
-                    renderer.write_line("not in a git worktree", c_error())?;
-                    return Ok(());
-                }
-            };
-            let target = if parts.len() >= 2 {
-                parts[1].trim().to_string()
-            } else {
-                match crate::extras::git_worktree::default_branch(&info.main_repo_path) {
-                    Some(b) => b,
-                    None => {
-                        renderer.write_line(
-                            "no target branch specified and couldn't detect main/master",
-                            c_error(),
-                        )?;
-                        return Ok(());
-                    }
-                }
-            };
-            let repo_name = crate::extras::git_worktree::repo_name(&info.main_repo_path);
-            let main_path = info.main_repo_path.display();
-            let wt_path = info.worktree_path.display();
-            renderer.write_line(
-                &format!(
-                    "merging '{}' into '{}' in {}...",
-                    info.branch, target, repo_name
-                ),
-                c_agent(),
-            )?;
-            return Err(anyhow::anyhow!(
-                "DEFER_WT_MERGE:{}:{}:{}:{}:{}",
-                info.branch,
-                target,
-                main_path,
-                wt_path,
-                repo_name
-            ));
-        }
-        #[cfg(feature = "git-worktree")]
-        "/wt-exit" => {
-            let info = match crate::extras::git_worktree::detect() {
-                Some(i) => i,
-                None => {
-                    renderer.write_line("not in a git worktree", c_error())?;
-                    return Ok(());
-                }
-            };
-            // EXT-9: refuse to exit a dirty worktree without an
-            // explicit `--force`. Otherwise pending edits / untracked
-            // files get silently stranded in the worktree directory
-            // while the user is dropped back into the main repo and
-            // assumes the work followed them. `git status --porcelain`
-            // emits a non-empty line per dirty path; any output =
-            // refuse.
-            let force = parts.iter().skip(1).any(|p| *p == "--force" || *p == "-f");
-            if !force {
-                let status = std::process::Command::new("git")
-                    .args(["status", "--porcelain"])
-                    .output();
-                match status {
-                    Ok(out) if out.status.success() && !out.stdout.is_empty() => {
-                        let dirty = String::from_utf8_lossy(&out.stdout);
-                        let line_count = dirty.lines().count();
-                        renderer.write_line(
-                            &format!(
-                                "worktree is dirty ({} uncommitted change{}); refusing to exit. Commit/stash first, or run `/wt-exit --force` to leave it stranded.",
-                                line_count,
-                                if line_count == 1 { "" } else { "s" },
-                            ),
-                            c_error(),
-                        )?;
-                        return Ok(());
-                    }
-                    Ok(_) => {} // clean tree
-                    Err(e) => {
-                        renderer.write_line(
-                            &format!(
-                                "could not run `git status` to check worktree state ({}); refusing to exit. Pass `--force` to override.",
-                                e
-                            ),
-                            c_error(),
-                        )?;
-                        return Ok(());
-                    }
-                }
-            }
-            let main_path = info.main_repo_path.display();
-            renderer.write_line(
-                &format!("returning to main repo at {}", main_path),
-                c_agent(),
-            )?;
-            return Err(anyhow::anyhow!(
-                "DEFER_WT_EXIT:{}:{}",
-                main_path,
-                info.worktree_path.display()
-            ));
-        }
-        "/regen-prompts" => match crate::context::prompts::regen() {
-            Ok(()) => {
-                context.prompts = crate::context::prompts::load();
-                // The active prompt's content may have changed on
-                // disk during regen. Re-bind `context.current_prompt`
-                // to the freshly-loaded body and rebuild the agent
-                // so the new preamble takes effect on the next turn
-                // without the user having to `/prompt <name>` again.
-                if let Some(name) = context.current_prompt_name.clone()
-                    && let Some(p) = context.prompts.get(&name)
-                {
-                    context.current_prompt = Some(p.body.clone());
-                    context.current_prompt_deny_tools = p.deny_tools.clone();
-                    crate::permission::apply_prompt_deny(
-                        permission,
-                        &context.current_prompt_deny_tools,
-                    );
-                }
-                let model = client.completion_model(session.model.to_string());
-                *agent = crate::provider::build_agent(
-                    model,
-                    cli,
-                    cfg,
-                    context,
-                    permission.clone(),
-                    ask_tx.clone(),
-                    question_tx.clone(),
-                    plan_tx.clone(),
-                    bg_store.clone(),
-                    #[cfg(feature = "lsp")]
-                    lsp_manager.cloned(),
-                    sandbox.clone(),
-                    #[cfg(feature = "mcp")]
-                    mcp_manager,
-                    #[cfg(feature = "semantic")]
-                    semantic_manager,
-                )
-                .await;
-                renderer.write_line(
-                    "default prompts regenerated; agent rebuilt with refreshed prompt",
-                    c_agent(),
-                )?;
-            }
-            Err(e) => {
-                renderer.write_line(&format!("failed to regenerate prompts: {}", e), c_error())?;
-            }
-        },
-        "/quit" => {
-            *is_running = false;
-            return Err(std::io::Error::new(std::io::ErrorKind::Interrupted, "quit").into());
-        }
-        // dirge-ov2 Phase C: `/tasks` LISTS chat windows but does not
-        // switch. Switching needs to save/load the UI loop's
-        // per-chat streaming state (response_buf, last_tool_name,
-        // etc.) which the slash handler doesn't have access to —
-        // switching from here would skip the save/load and corrupt
-        // the active chat's mid-stream context. Users switch via
-        // Ctrl-N / Ctrl-P / Ctrl-X (which DO go through the save/
-        // load path in the UI loop).
-        "/tasks" => {
-            let names = renderer.chat_names();
-            if names.len() <= 1 {
-                renderer.write_line(
-                    "no subagent chats yet — spawn one via the `task` tool.",
-                    c_result(),
-                )?;
-            } else {
-                renderer.write_line("chat windows:", c_result())?;
-                let active = renderer.active_chat();
-                for (i, name) in names.iter().enumerate() {
-                    let marker = if i == active { "→" } else { " " };
-                    renderer.write_line(&format!("  {} [{}] {}", marker, i, name), c_result())?;
-                }
-                renderer.write_line("  (Ctrl-N / Ctrl-P / Ctrl-X to switch)", theme::dim())?;
-            }
-        }
-        "/clear" => {
-            session.messages.clear();
-            session.total_estimated_tokens = 0;
-            session.compactions.clear();
-            // Drop branch state too so a fresh /clear truly starts over.
-            // Without this, /tree would still show the cleared messages
-            // as a dead-end branch.
-            session.message_store.clear();
-            session.tree.entries.clear();
-            session.tree.leaf_id = None;
-            crate::agent::tools::modified::clear_modified();
-            render_session(renderer, session, cli, cfg, context)?;
-        }
-        "/tree" => {
-            // No-arg: print an ASCII view of the tree.
-            // <id-prefix>: switch the active branch to the leaf matching
-            //              the given id prefix (no need to type full UUID).
-            session.ensure_tree_initialized();
-            session.ensure_message_store_initialized();
-            let arg = parts.get(1).copied().unwrap_or("").trim();
-            if arg.is_empty() {
-                if session.tree.entries.is_empty() {
-                    renderer.write_line("(empty session)", c_agent())?;
-                } else {
-                    for line in tree::render_tree(session) {
-                        renderer.write_line(&line, c_result())?;
-                    }
-                }
-            } else {
-                match tree::resolve_id_prefix(session, arg) {
-                    Ok(id) => {
-                        if let Err(e) = session.switch_to_leaf(&id) {
-                            renderer.write_line(&format!("switch failed: {}", e), c_error())?;
-                        } else {
-                            render_session(renderer, session, cli, cfg, context)?;
-                            renderer.write_line(
-                                &format!("switched to leaf {}", short_id(&id)),
-                                c_agent(),
-                            )?;
-                        }
-                    }
-                    Err(e) => renderer.write_line(&format!("/tree: {}", e), c_error())?,
-                }
-            }
-        }
-        "/fork" => {
-            // /fork [id-prefix] — branch off from the parent of the
-            // chosen message, and pop the original prompt back into
-            // the editor so the user can re-edit and retry.
-            // Default target: the most recent user message on the
-            // current branch (i.e. "redo last prompt").
-            session.ensure_tree_initialized();
-            session.ensure_message_store_initialized();
-            let arg = parts.get(1).copied().unwrap_or("").trim();
-            let target_id = if arg.is_empty() {
-                // Default to the last User message on the current path.
-                let last_user = session
-                    .messages
-                    .iter()
-                    .rev()
-                    .find(|m| m.role == MessageRole::User)
-                    .map(|m| m.id.clone());
-                match last_user {
-                    Some(id) => Ok(id),
-                    None => Err("no user message on current branch".to_string()),
-                }
-            } else {
-                tree::resolve_id_prefix(session, arg)
-            };
-            match target_id {
-                Ok(id) => match session.fork_at(&id) {
-                    Ok(original) => {
-                        input.set_text(&original.content);
-                        render_session(renderer, session, cli, cfg, context)?;
-                        renderer.write_line(
-                            &format!(
-                                "forked at {} — original prompt restored to editor",
-                                short_id(&id)
-                            ),
-                            c_agent(),
-                        )?;
-                    }
-                    Err(e) => renderer.write_line(&format!("/fork: {}", e), c_error())?,
-                },
-                Err(e) => renderer.write_line(&format!("/fork: {}", e), c_error())?,
-            }
-        }
-        "/clone" => {
-            // /clone <id-prefix> — make the chosen entry the leaf
-            // without restoring its content into the editor. Useful
-            // for jumping to a labeled bookmark or comparing branches.
-            session.ensure_tree_initialized();
-            session.ensure_message_store_initialized();
-            let arg = parts.get(1).copied().unwrap_or("").trim();
-            if arg.is_empty() {
-                renderer.write_line("usage: /clone <id-prefix>", c_error())?;
-            } else {
-                match tree::resolve_id_prefix(session, arg) {
-                    Ok(id) => match session.clone_at(&id) {
-                        Ok(()) => {
-                            render_session(renderer, session, cli, cfg, context)?;
-                            renderer.write_line(
-                                &format!("cloned path through {}", short_id(&id)),
-                                c_agent(),
-                            )?;
-                        }
-                        Err(e) => renderer.write_line(&format!("/clone: {}", e), c_error())?,
-                    },
-                    Err(e) => renderer.write_line(&format!("/clone: {}", e), c_error())?,
-                }
-            }
-        }
-        "/panel" => {
-            use crate::ui::renderer::PanelMode;
-            let arg = parts.get(1).copied().unwrap_or("").trim();
-            let new_mode = match arg {
-                "" => None,
-                "on" => Some(PanelMode::On),
-                "off" => Some(PanelMode::Off),
-                "auto" => Some(PanelMode::Auto),
-                other => {
-                    renderer.write_line(
-                        &format!("unknown /panel mode '{}' (use on|off|auto)", other),
-                        c_error(),
-                    )?;
-                    return Ok(());
-                }
-            };
-            if let Some(mode) = new_mode {
-                renderer.set_panel_mode(mode);
-                // Force a full repaint so layout / clipping recomputes at
-                // the new width immediately, not on next event.
-                renderer.render_viewport()?;
-            }
-            let current = renderer.panel_mode();
-            let visible = renderer.panel_visible();
-            renderer.write_line(
-                &format!(
-                    "panel mode: {:?} (currently {})",
-                    current,
-                    if visible { "shown" } else { "hidden" }
-                ),
-                c_agent(),
-            )?;
-        }
-        "/btw" => {
-            let query = parts.get(1..).map(|p| p.join(" ")).unwrap_or_default();
-            if query.is_empty() {
-                renderer.write_line("usage: /btw <question>", c_error())?;
-            } else {
-                let model = client.completion_model(session.model.to_string());
-                renderer.write_line(&format!("btw: {}", query), Color::DarkGrey)?;
-                match model.btw_query(query).await {
-                    Ok(response) => {
-                        renderer.write_line("", Color::White)?;
-                        let max_width = renderer.line_width();
-                        let styled = crate::ui::markdown::markdown_to_styled(
-                            &response,
-                            max_width,
-                            c_agent(),
-                        );
-                        for span in styled {
-                            renderer.write(&span.text, span.color)?;
-                        }
-                        renderer.write_line("", Color::White)?;
-                    }
-                    Err(e) => {
-                        renderer.write_line(&format!("btw error: {}", e), c_error())?;
-                    }
-                }
-            }
-        }
-        "/cd" => {
-            // `splitn(3, ' ')` produces empty middle elements for
-            // consecutive spaces (`/cd   /tmp` → `["/cd", "", "/tmp"]`),
-            // which collapses to "cd home" — silent surprise.
-            // Re-derive the target from everything after the slash
-            // command using whitespace-aware splitting.
-            let raw_args = text.trim().strip_prefix("/cd").unwrap_or("").trim();
-            let target = raw_args;
-            let path = if target.is_empty() {
-                dirs::home_dir().unwrap_or_default()
-            } else if let Some(rest) = target.strip_prefix('~') {
-                let mut home = dirs::home_dir().unwrap_or_default();
-                home.push(rest.trim_start_matches('/'));
-                home
-            } else {
-                std::path::PathBuf::from(target)
-            };
-            match std::env::set_current_dir(&path) {
-                Ok(()) => {
-                    let canonical = std::fs::canonicalize(&path).unwrap_or(path);
-                    session.working_dir =
-                        compact_str::CompactString::new(canonical.to_string_lossy());
-                    if let Some(perm) = permission {
-                        if let Ok(mut guard) = perm.lock() {
-                            guard.set_working_dir(&session.working_dir);
-                        }
-                    }
-                    context.reload();
-                    let model = client.completion_model(session.model.to_string());
-                    *agent = crate::provider::build_agent(
-                        model,
-                        cli,
-                        cfg,
-                        context,
-                        permission.clone(),
-                        ask_tx.clone(),
-                        question_tx.clone(),
-                        plan_tx.clone(),
-                        bg_store.clone(),
-                        #[cfg(feature = "lsp")]
-                        lsp_manager.cloned(),
-                        sandbox.clone(),
-                        #[cfg(feature = "mcp")]
-                        mcp_manager,
-                        #[cfg(feature = "semantic")]
-                        semantic_manager,
-                    )
-                    .await;
-                    render_session(renderer, session, cli, cfg, context)?;
-                    renderer.write_line(
-                        &format!("changed directory to {}", session.working_dir),
-                        c_agent(),
-                    )?;
-                }
-                Err(e) => {
-                    renderer.write_line(&format!("cd: {}", e), c_error())?;
-                }
-            }
-        }
-        "/undo" => {
-            let outcome = undo_last(session);
-            if outcome.removed > 0 {
-                render_session(renderer, session, cli, cfg, context)?;
-                renderer.write_line(
-                    &format!("removed {} message(s)", outcome.removed),
-                    c_agent(),
-                )?;
-                // SESS-13: rewinding the transcript does NOT undo
-                // any file writes, bash commands, MCP calls, or
-                // other side effects executed during the popped
-                // turn — only the LLM's view of history. Surface
-                // this so the user doesn't assume the on-disk
-                // state was reverted.
-                if outcome.had_tool_calls {
-                    renderer.write_line(
-                        "warning: tool side effects (file writes, bash, MCP) were NOT reverted",
-                        c_error(),
-                    )?;
-                }
-            } else {
-                renderer.write_line("nothing to undo", c_agent())?;
-            }
-        }
-        "/retry" => {
-            let last_user = session
-                .messages
-                .iter()
-                .rev()
-                .find(|m| m.role == MessageRole::User)
-                .cloned();
-            match last_user {
-                Some(msg) => {
-                    // Pop messages until we've removed the last user
-                    // message itself — covers assistant replies AND
-                    // system messages (compress summaries, error
-                    // notices) that landed between the user prompt
-                    // and the current state. `undo_last` only handles
-                    // the Assistant/User pair pattern; using it here
-                    // would leave a trailing System message in the
-                    // session and the agent would see it on retry.
-                    //
-                    // Bound the loop at the current message count so
-                    // a corrupt session that somehow lacks user msgs
-                    // (despite the `last_user.is_some()` check above)
-                    // can't infinite-loop on `pop` returning None.
-                    let mut guard = session.messages.len();
-                    while let Some(last) = session.messages.last() {
-                        let was_user = last.role == MessageRole::User;
-                        session.pop_last_message();
-                        if was_user {
-                            break;
-                        }
-                        guard = guard.saturating_sub(1);
-                        if guard == 0 {
-                            break;
-                        }
-                    }
-                    input.buffer = msg.content.clone();
-                    input.cursor = msg.content.len();
-                    render_session(renderer, session, cli, cfg, context)?;
-                    renderer.write_line("edit last message and press Enter to retry", c_agent())?;
-                }
-                None => {
-                    renderer.write_line("no previous message to retry", c_agent())?;
-                }
-            }
-        }
-        "/allow" => {
-            // Phase 5: CRUD for the session permission allowlist.
-            // Subcommands: list (default), add <tool> <pattern>,
-            // remove <idx>, clear. Without this, users can only
-            // create allowlist entries via the interactive "(a)
-            // allow always" prompt; there's no way to inspect or
-            // undo a bad grant short of editing the session JSON.
-            let sub = parts.get(1).copied().unwrap_or("list");
-            let perm = match permission {
-                Some(p) => p,
-                None => {
-                    renderer.write_line(
-                        "permission system unavailable (--no-tools mode?)",
-                        c_error(),
-                    )?;
-                    return Ok(());
-                }
-            };
-            match sub {
-                "list" => {
-                    let entries = {
-                        let guard = perm.lock().unwrap_or_else(|e| e.into_inner());
-                        guard.allowlist_entries()
-                    };
-                    if entries.is_empty() {
-                        renderer.write_line(
-                            "session allowlist is empty (use '(a) allow always' in a permission prompt to add entries)",
-                            c_agent(),
-                        )?;
-                    } else {
-                        renderer.write_line(
-                            &format!("session allowlist ({} entries):", entries.len()),
-                            c_agent(),
-                        )?;
-                        for (i, (tool, pat)) in entries.iter().enumerate() {
-                            renderer
-                                .write_line(&format!("  [{}] {} {}", i, tool, pat), c_result())?;
-                        }
-                        renderer.write_line(
-                            "use '/allow remove <idx>' to drop a single entry; '/allow clear' to drop all",
-                            theme::dim(),
-                        )?;
-                    }
-                }
-                "add" => {
-                    // `/allow add <tool> <pattern>` — third part is
-                    // the tool, rest is the pattern (may contain
-                    // spaces, so re-derive from raw text).
-                    let raw_args = text.trim().strip_prefix("/allow").unwrap_or("").trim();
-                    let rest = raw_args.strip_prefix("add").unwrap_or("").trim();
-                    let mut it = rest.splitn(2, char::is_whitespace);
-                    let tool = it.next().unwrap_or("");
-                    let pattern = it.next().unwrap_or("").trim();
-                    // Whitelist accepted tool names so a typo
-                    // (`/allow add bsah ...`) doesn't silently
-                    // create a permanently-inert rule the user
-                    // can't figure out. Matches the schema used
-                    // by `PermissionConfig`.
-                    const KNOWN_PERM_TOOLS: &[&str] = &[
-                        "bash",
-                        "read",
-                        "write",
-                        "edit",
-                        "grep",
-                        "find_files",
-                        "glob",
-                        "list_dir",
-                        "write_todo_list",
-                        "apply_patch",
-                        "lsp",
-                        "question",
-                        "webfetch",
-                        "websearch",
-                        "task",
-                        "task_status",
-                        "memory",
-                        "skill",
-                        "list_symbols",
-                        "get_symbol_body",
-                        "find_definition",
-                        "find_callers",
-                        "find_callees",
-                        "mcp_tool",
-                    ];
-                    if tool.is_empty() || pattern.is_empty() {
-                        renderer.write_line(
-                            "usage: /allow add <tool> <pattern>  (e.g. /allow add bash 'cargo *')",
-                            c_error(),
-                        )?;
-                    } else if !KNOWN_PERM_TOOLS.contains(&tool) {
-                        renderer.write_line(
-                            &format!(
-                                "unknown tool {:?}. Valid: {}",
-                                tool,
-                                KNOWN_PERM_TOOLS.join(", "),
-                            ),
-                            c_error(),
-                        )?;
-                    } else {
-                        {
-                            let mut guard = perm.lock().unwrap_or_else(|e| e.into_inner());
-                            guard.add_session_allowlist(tool.to_string(), pattern);
-                        }
-                        // Mirror into session.permission_allowlist
-                        // so the entry persists across save/load.
-                        let entry = crate::session::PermissionAllowEntry {
-                            tool: tool.to_string(),
-                            pattern: pattern.to_string(),
-                        };
-                        // Dedup at the session level too — checker
-                        // dedupes but we want save() to write a
-                        // clean list.
-                        if !session
-                            .permission_allowlist
-                            .iter()
-                            .any(|e| e.tool == entry.tool && e.pattern == entry.pattern)
-                        {
-                            session.permission_allowlist.push(entry);
-                        }
-                        renderer.write_line(&format!("added: {} {}", tool, pattern), c_agent())?;
-                    }
-                }
-                "remove" => {
-                    let idx_str = parts.get(2).copied().unwrap_or("");
-                    let idx: usize = match idx_str.parse() {
-                        Ok(n) => n,
-                        Err(_) => {
-                            renderer.write_line(
-                                "usage: /allow remove <idx>  (run /allow list to see indices)",
-                                c_error(),
-                            )?;
-                            return Ok(());
-                        }
-                    };
-                    let removed = {
-                        let mut guard = perm.lock().unwrap_or_else(|e| e.into_inner());
-                        guard.remove_session_allowlist_at(idx)
-                    };
-                    match removed {
-                        Some((tool, pat)) => {
-                            // Mirror removal into the session
-                            // allowlist too.
-                            session
-                                .permission_allowlist
-                                .retain(|e| !(e.tool == tool && e.pattern == pat));
-                            renderer.write_line(
-                                &format!("removed [{}]: {} {}", idx, tool, pat),
-                                c_agent(),
-                            )?;
-                        }
-                        None => {
-                            renderer.write_line(
-                                &format!("no allowlist entry at index {}", idx),
-                                c_error(),
-                            )?;
-                        }
-                    }
-                }
-                "clear" => {
-                    {
-                        let mut guard = perm.lock().unwrap_or_else(|e| e.into_inner());
-                        guard.clear_session_allowlist();
-                    }
-                    session.permission_allowlist.clear();
-                    renderer.write_line("session allowlist cleared", c_agent())?;
-                }
-                other => {
-                    renderer.write_line(
-                        &format!(
-                            "unknown /allow subcommand {:?}; try: list, add, remove, clear",
-                            other,
-                        ),
-                        c_error(),
-                    )?;
-                }
-            }
-        }
-        "/help" => {
-            renderer.write_line("commands:", c_agent())?;
-            renderer.write_line("  /model [name]          show or switch model", c_result())?;
-            renderer.write_line("  /sessions              list recent sessions", c_result())?;
-            renderer.write_line(
-                "  /sessions <id>         load a session (by ID prefix)",
-                c_result(),
-            )?;
-            renderer.write_line("  /sessions delete <id>  delete a session", c_result())?;
-            renderer.write_line(
-                "  /reasoning             toggle reasoning visibility",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /mode                  show/change security mode",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /mode <mode>           set mode (standard|restrictive|accept|yolo)",
-                c_result(),
-            )?;
-            #[cfg(feature = "mcp")]
-            {
-                let _ = renderer.write_line(
-                    "  /mcp                   list MCP servers and tools",
-                    c_result(),
-                );
-                let _ = renderer.write_line(
-                    "  /mcp <server>          list tools of an MCP server",
-                    c_result(),
-                );
-            }
-            renderer.write_line(
-                "  /clear                 clear screen + reset tree",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /tree                  show the session tree (use /tree <id-prefix> to switch branches)",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /fork [id-prefix]      branch off at the chosen message (default: last user message)",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /clone <id-prefix>     switch to the branch ending at the chosen entry",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /panel [on|off|auto]   toggle right-hand info panel",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /cd [path]             change working directory",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /btw <question>        ask a quick question (no tools, doesn't affect session)",
-                c_result(),
-            )?;
-            renderer.write_line("  /undo                  undo last exchange", c_result())?;
-            renderer.write_line("  /retry                 retry last prompt", c_result())?;
-            renderer.write_line(
-                "  /compress [/compact]   compress conversation history",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /compress [focus]      compress; focus text guides what to preserve",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /toggle <feat> [on|off] toggle a feature (e.g. /toggle todo)",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /allow list            list session allowlist entries",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /allow add <tool> <pat> add an allowlist entry",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /allow remove <idx>    drop one allowlist entry",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /allow clear           drop all allowlist entries",
-                c_result(),
-            )?;
-            #[cfg(feature = "loop")]
-            {
-                let _ = renderer.write_line(
-                    "  /loop [prompt]         start iterative coding loop",
-                    c_result(),
-                );
-                let _ = renderer.write_line("  /loop stop             stop the loop", c_result());
-            }
-            #[cfg(not(feature = "loop"))]
-            {
-                let _ = renderer.write_line(
-                    "  /loop [prompt]         start iterative coding loop (req. 'loop' feature)",
-                    c_result(),
-                );
-            }
-            renderer.write_line(
-                "  /prompt                list available prompts (plan|code)",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /prompt <name>         activate a prompt by name",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /prompt default        activate the 'default' prompt if installed,",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "                         otherwise clears the active prompt",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  /regen-prompts        restore built-in prompts to global dir",
-                c_result(),
-            )?;
-            #[cfg(feature = "git-worktree")]
-            {
-                let _ = renderer.write_line(
-                    "  /worktree <name>       create a git worktree on <name> branch and cd into it",
-                    c_result(),
-                );
-                let _ = renderer.write_line(
-                    "  /wt-merge [branch]     merge worktree branch into [branch] (default: main/master)",
-                    c_result(),
-                );
-                let _ = renderer.write_line(
-                    "  /wt-exit               exit worktree and return to main repo",
-                    c_result(),
-                );
-            }
-            renderer.write_line("  /quit                  exit dirge", c_result())?;
-            renderer.write_line("  /help                  show this message", c_result())?;
-            renderer.write_line("", c_agent())?;
-            renderer.write_line("keys:", c_agent())?;
-            renderer.write_line("  PgUp/PgDn / wheel      scroll chat history", c_result())?;
-            renderer.write_line("  Home/End               jump to top/bottom", c_result())?;
-            renderer.write_line(
-                "  @<query>               file picker (Tab/Enter select, Esc cancel)",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  drag                  select text; mouse-up copies to clipboard",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  y / Esc (while selecting)  copy selection / clear without copy",
-                c_result(),
-            )?;
-            renderer.write_line("  Ctrl+R                 toggle reasoning", c_result())?;
-            renderer.write_line("  Ctrl+C / Ctrl+D        interrupt/quit", c_result())?;
-            renderer.write_line(
-                "  Ctrl+N / Ctrl+P        next / previous chat (subagent windows)",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  Ctrl+X / /tasks        cycle through chat windows",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  Alt+X                  drop last queued interjection",
-                c_result(),
-            )?;
-            renderer.write_line(
-                "  (type while agent runs to queue a follow-up message)",
-                c_result(),
-            )?;
-            renderer.write_line("  PgUp/PgDn / wheel      scroll chat history", c_result())?;
-
-            // Plugin-registered commands, if any. Listed last so they sit
-            // visually after the built-ins and the keybindings.
-            #[cfg(feature = "plugin")]
-            if let Some(pm_arc) = crate::plugin::hook::global() {
-                let cmds = {
-                    let mut mgr = pm_arc.lock().unwrap_or_else(|e| e.into_inner());
-                    mgr.list_commands()
-                };
-                if !cmds.is_empty() {
-                    renderer.write_line("", c_agent())?;
-                    renderer.write_line("plugin commands:", c_agent())?;
-                    for (cmd, handler) in cmds {
-                        renderer
-                            .write_line(&format!("  /{:<20} -> {}", cmd, handler), c_result())?;
-                    }
-                }
-            }
-        }
-        _ => {
-            // If `slash_command_names()` advertised this command
-            // but no match arm above caught it, the lists drifted
-            // (added to the canonical list without wiring up the
-            // dispatch). Emit a loud error here rather than falling
-            // through to plugin lookup / "unknown command", so the
-            // mistake is obvious in dev/test rather than silently
-            // shadowed by either path. Plugin commands by
-            // convention don't have a leading `/` in the canonical
-            // list, so this won't false-fire on them.
-            if is_known_slash_command(parts[0]) {
-                renderer.write_line(
-                    &format!(
-                        "internal error: {} is listed in slash_command_names() but has no dispatch arm in handle_slash — wire it up or remove from the list",
-                        parts[0]
-                    ),
-                    c_error(),
-                )?;
-                return Ok(());
-            }
-
-            // Fall through to plugin-registered commands. The process-global
-            // PluginManager is the same one HookedToolDyn uses, so we don't
-            // need to thread an Arc through handle_slash's already long
-            // parameter list.
-            #[cfg(feature = "plugin")]
-            if let Some(pm_arc) = crate::plugin::hook::global() {
-                let cmd = parts[0].trim_start_matches('/');
-                let args = parts.get(1..).map(|p| p.join(" ")).unwrap_or_default();
-                let handler = {
-                    let mut mgr = pm_arc.lock().unwrap_or_else(|e| e.into_inner());
-                    mgr.list_commands()
-                        .into_iter()
-                        .find(|(name, _)| name == cmd)
-                        .map(|(_, h)| h)
-                };
-                if let Some(handler_fn) = handler {
-                    let result = {
-                        let mut mgr = pm_arc.lock().unwrap_or_else(|e| e.into_inner());
-                        mgr.invoke_command(&handler_fn, &args)
-                    };
-                    match result {
-                        Ok(Some(text)) => {
-                            // Strip ANSI escapes from plugin output to
-                            // prevent repaint/screen-manipulation attacks.
-                            let safe = crate::ui::ansi::strip_escapes(
-                                &text,
-                                crate::ui::ansi::StripPolicy::KEEP_NEWLINE,
-                            );
-                            for line in safe.lines() {
-                                renderer.write_line(line, c_agent())?;
-                            }
-                        }
-                        Ok(None) => {
-                            // Handler ran cleanly but had nothing to say — no-op.
-                        }
-                        Err(e) => {
-                            renderer.write_line(
-                                &format!("[plugin] {} failed: {}", cmd, e),
-                                c_error(),
-                            )?;
-                        }
-                    }
-                    return Ok(());
-                }
-            }
-            renderer.write_line(
-                &format!("unknown command: {} (try /help)", parts[0]),
-                c_error(),
-            )?;
-        }
-    }
-    Ok(())
-}
-
-/// Result of a slash-command tab completion.
-#[allow(dead_code)]
-pub struct CompletionResult {
-    pub new_buffer: String,
-    pub new_cursor: usize,
-    /// The full sorted command list and the index of the currently-selected
-    /// command, so the renderer can show a preview of upcoming items.
-    pub all_commands: Vec<&'static str>,
-    pub current_index: usize,
-}
-
-/// Try to complete the slash command at `cursor` in `buffer`.
-/// Returns `Some(CompletionResult)` if completion was possible.
-/// Tab cycles through matching commands; when narrowed to one match,
-/// subsequent tabs cycle through ALL commands so the user can keep browsing.
-#[cfg(feature = "experimental-ui-tab-slash")]
-pub fn try_complete(buffer: &str, cursor: usize) -> Option<CompletionResult> {
-    if !buffer.starts_with('/') {
-        return None;
-    }
-
-    let cursor = cursor.min(buffer.len());
-    // Find the bounds of the FIRST word in the buffer (the command
-    // name). Previously the replacement range was
-    // `[word_start..cursor]`, which corrupted the buffer when the
-    // cursor sat mid-word: e.g. Tab with cursor at byte 2 of `/mod`
-    // produced `/mcpod` (replacement appended to the tail after the
-    // cursor). Anchor the replacement to the whole word boundary
-    // instead — the cursor can land anywhere inside the command and
-    // the result is still well-formed.
-    let word_start = 0usize;
-    let word_end = buffer.find(char::is_whitespace).unwrap_or(buffer.len());
-    let current_word = &buffer[word_start..word_end];
-
-    // Only complete when the cursor is inside (or just after) the
-    // first word. Cursor past the first whitespace means the user is
-    // typing args, not a command name.
-    if cursor > word_end {
-        return None;
-    }
-
-    let all_commands = builtin_commands();
-    let matching: Vec<&str> = all_commands
-        .iter()
-        .filter(|c| c.starts_with(current_word))
-        .copied()
-        .collect();
-
-    if matching.is_empty() {
-        return None;
-    }
-
-    // Once the current word is an exact command name, cycle through ALL
-    // commands so the user can keep browsing.  Otherwise stay within the
-    // matching prefix subset (e.g. /mod → /mode, /model → repeats).
-    let is_exact = all_commands.contains(&current_word);
-
-    let (replacement, current_index) = if is_exact {
-        let all_idx = all_commands.iter().position(|c| *c == current_word);
-        let next_idx = match all_idx {
-            Some(i) => (i + 1) % all_commands.len(),
-            None => 0,
-        };
-        (all_commands[next_idx], next_idx)
-    } else {
-        let current_idx = matching.iter().position(|c| *c == current_word);
-        let next_idx = match current_idx {
-            Some(i) => (i + 1) % matching.len(),
-            None => 0,
-        };
-        let cmd = matching[next_idx];
-        let all_idx = all_commands.iter().position(|c| *c == cmd).unwrap_or(0);
-        (cmd, all_idx)
-    };
-    // Build the new buffer: prefix (everything before the word —
-    // always empty here since word_start==0) + replacement + tail
-    // (everything from word_end onward). Cursor lands at the end of
-    // the replacement so the user can immediately type args.
-    let mut new_buffer = String::with_capacity(replacement.len() + buffer.len() - word_end);
-    new_buffer.push_str(replacement);
-    new_buffer.push_str(&buffer[word_end..]);
-    let new_cursor = replacement.len();
-    Some(CompletionResult {
-        new_buffer,
-        new_cursor,
-        all_commands,
-        current_index,
-    })
-}
-
-/// Canonical list of built-in slash commands. **Single source of
-/// truth** consulted by:
-///   * `builtin_commands()`     — tab completion (feature-gated)
-///   * `is_known_slash_command` — handle_slash's "internal error"
-///     vs "unknown command" branching in the default arm
-///
-/// **When you add a new slash command to `handle_slash`'s match
-/// arms, add it here too.** A drift in the other direction (listed
-/// here but no match arm in `handle_slash`) surfaces at runtime as
-/// an explicit `internal error: known command X reached default
-/// arm` so the mistake is loud rather than silent.
-///
-/// Always-compiled (not feature-gated) because `handle_slash`'s
-/// default arm consults it regardless of the tab-completion
-/// feature.
-pub fn slash_command_names() -> Vec<&'static str> {
-    let mut cmds = vec![
-        "/allow",
-        "/btw",
-        "/cd",
-        "/clear",
-        "/clone",
-        "/compact",
-        "/compress",
-        "/fork",
-        "/help",
-        "/mode",
-        "/model",
-        "/panel",
-        "/prompt",
-        "/quit",
-        "/reasoning",
-        "/regen-prompts",
-        "/retry",
-        "/sessions",
-        "/tasks",
-        "/toggle",
-        "/tree",
-        "/undo",
-    ];
-    #[cfg(feature = "git-worktree")]
-    {
-        cmds.push("/worktree");
-        cmds.push("/wt-exit");
-        cmds.push("/wt-merge");
-    }
-    #[cfg(feature = "mcp")]
-    cmds.push("/mcp");
-    #[cfg(feature = "loop")]
-    cmds.push("/loop");
-    cmds.sort_unstable();
-    cmds
-}
-
-/// Returns true if `name` (with leading `/`) is a built-in slash
-/// command. Used by `handle_slash`'s default arm to distinguish
-/// "command name we should have dispatched but didn't" (internal
-/// error) from "command name we don't know about" (plugin fallback /
-/// unknown).
-pub fn is_known_slash_command(name: &str) -> bool {
-    slash_command_names().iter().any(|c| *c == name)
-}
-
-/// Returns all built-in slash commands (with leading `/`), sorted alphabetically.
-#[cfg(feature = "experimental-ui-tab-slash")]
-pub fn builtin_commands() -> Vec<&'static str> {
-    slash_command_names()
-}
-
-/// Format a completion preview string showing upcoming commands.
-/// Returns an empty string when `cr` is `None`. The result is shaped
-/// to fit within `avail_w` display cells (after the continuation
-/// prompt), showing as many upcoming command names as will fit.
-#[cfg(feature = "experimental-ui-tab-slash")]
-pub fn format_completion_preview(cr: Option<&CompletionResult>, avail_w: usize) -> String {
-    let cr = match cr {
-        Some(c) => c,
-        None => return String::new(),
-    };
-    if cr.all_commands.is_empty() || avail_w < 4 {
-        return String::new();
-    }
-    let all = &cr.all_commands;
-    let start = (cr.current_index + 1) % all.len();
-    let mut result = String::new();
-    for i in 0..all.len() {
-        let cmd = all[(start + i) % all.len()];
-        let candidate = if result.is_empty() {
-            cmd.to_string()
-        } else {
-            format!("{result}  {cmd}")
-        };
-        use unicode_width::UnicodeWidthStr;
-        if UnicodeWidthStr::width(candidate.as_str()) > avail_w {
-            break;
-        }
-        result = candidate;
-    }
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::session::{Session, SessionMessage};
-
-    fn msg(role: MessageRole, content: &str) -> SessionMessage {
-        // Re-use Session::add_message to get a real msg with id/timestamp.
-        let mut s = Session::new("p", "m", 0);
-        s.add_message(role, content);
-        s.messages.pop().unwrap()
-    }
-
-    /// F3: when the reverse-scan lands on an Assistant message, the
-    /// helper advances to the next User so the kept tail begins
-    /// with a User message (provider-required role sequence after
-    /// the System summary).
-    #[test]
-    fn align_cut_advances_past_assistant_to_next_user() {
-        let messages = vec![
-            msg(MessageRole::User, "u0"),
-            msg(MessageRole::Assistant, "a0"),
-            msg(MessageRole::User, "u1"),
-            msg(MessageRole::Assistant, "a1"), // reverse-scan landed here (cut_idx=3)
-            msg(MessageRole::User, "u2"),
-            msg(MessageRole::Assistant, "a2"),
-        ];
-        // Initial cut at idx=3 (Assistant). Should advance to 4 (User).
-        assert_eq!(align_cut_to_user_boundary(&messages, 3), 4);
-    }
-
-    /// User-boundary cut is unchanged.
-    #[test]
-    fn align_cut_idempotent_when_already_on_user() {
-        let messages = vec![
-            msg(MessageRole::User, "u0"),
-            msg(MessageRole::Assistant, "a0"),
-            msg(MessageRole::User, "u1"),
-        ];
-        assert_eq!(align_cut_to_user_boundary(&messages, 2), 2);
-        assert_eq!(align_cut_to_user_boundary(&messages, 0), 0);
-    }
-
-    /// Cut past end of array stays past end.
-    #[test]
-    fn align_cut_past_end_clamps() {
-        let messages = vec![msg(MessageRole::User, "u0")];
-        assert_eq!(align_cut_to_user_boundary(&messages, 1), 1);
-        assert_eq!(align_cut_to_user_boundary(&messages, 5), 5);
-    }
-
-    /// No user in the tail (e.g. only system+assistant remain after
-    /// the cut). Helper returns `messages.len()`, which is the
-    /// "nothing to compress (no clean boundary)" case.
-    #[test]
-    fn align_cut_returns_end_when_no_user_in_tail() {
-        let messages = vec![
-            msg(MessageRole::User, "u0"),
-            msg(MessageRole::Assistant, "a0"),
-            msg(MessageRole::System, "system note"),
-            msg(MessageRole::Assistant, "a1"),
-        ];
-        // cut_idx=2 points at System; no User follows.
-        assert_eq!(align_cut_to_user_boundary(&messages, 2), messages.len());
-    }
-
-    /// A cut that lands on a System message (e.g. a prior summary)
-    /// also advances forward to the next User.
-    #[test]
-    fn align_cut_skips_system_to_user() {
-        let messages = vec![
-            msg(MessageRole::System, "prior summary"),
-            msg(MessageRole::User, "u0"),
-            msg(MessageRole::Assistant, "a0"),
-        ];
-        assert_eq!(align_cut_to_user_boundary(&messages, 0), 1);
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn no_completion_without_slash() {
-        assert!(try_complete("hello", 5).is_none());
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn empty_buffer_returns_none() {
-        assert!(try_complete("", 0).is_none());
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn complete_partial_command() {
-        let r = try_complete("/mod", 4).unwrap();
-        assert_eq!(r.new_buffer, "/mode");
-        assert_eq!(r.new_cursor, 5);
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn cycles_between_partial_matches() {
-        let r = try_complete("/mod", 4).unwrap();
-        assert!(r.new_buffer.starts_with("/mod"));
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn cycles_beyond_single_match() {
-        let r1 = try_complete("/", 1).unwrap();
-        let r2 = try_complete(&r1.new_buffer, r1.new_cursor).unwrap();
-        assert_ne!(r1.new_buffer, r2.new_buffer);
-        assert!(!r2.new_buffer.is_empty());
-        assert!(r2.new_buffer.starts_with('/'));
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn cycles_from_full_command() {
-        let r = try_complete("/btw", 4).unwrap();
-        assert_ne!(r.new_buffer, "/btw");
-        assert!(r.new_buffer.starts_with('/'));
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn cycles_through_all_commands() {
-        let mut seen = std::collections::HashSet::new();
-        let mut buf = "/".to_string();
-        let mut cur = 1;
-        for _ in 0..100 {
-            let result = try_complete(&buf, cur);
-            if result.is_none() {
-                break;
-            }
-            let r = result.unwrap();
-            buf = r.new_buffer;
-            cur = r.new_cursor;
-            seen.insert(buf.clone());
-        }
-        let all = builtin_commands();
-        assert_eq!(
-            seen.len(),
-            all.len(),
-            "should cycle through all builtin commands"
-        );
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn unknown_command_returns_none() {
-        assert!(try_complete("/nonexistent", 12).is_none());
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn commands_are_sorted() {
-        let cmds = builtin_commands();
-        for pair in cmds.windows(2) {
-            assert!(
-                pair[0] <= pair[1],
-                "{} should be before {}",
-                pair[0],
-                pair[1]
-            );
-        }
-    }
-
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn preview_includes_upcoming_commands() {
-        let r = try_complete("/", 1).unwrap();
-        let all = &r.all_commands;
-        let cur = r.current_index;
-        let upcoming = &all[(cur + 1)..];
-        assert!(
-            !upcoming.is_empty(),
-            "should have commands after the current one"
-        );
-    }
-
-    // ============================================================
-    // Code-review B1 fix: cursor mid-word produces well-formed
-    // buffer
-    // ============================================================
-
-    /// Regression: cursor sitting at byte 2 of `/mod` previously
-    /// produced `/mcpod` (replacement appended to the tail AFTER
-    /// the cursor, with `od` left over from the original word).
-    /// The fix anchors replacement to the whole-word boundary
-    /// instead, so cursor position inside the command name doesn't
-    /// corrupt the buffer — the result is exactly one of the
-    /// matching commands, no Frankenstein.
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn complete_with_cursor_mid_word_produces_clean_buffer() {
-        // /mod, cursor at the `o` (byte 2). The new buffer must be
-        // exactly a candidate command — not the candidate +
-        // residual `od` from the source.
-        let r = try_complete("/mod", 2).unwrap();
-        let candidates = builtin_commands()
-            .into_iter()
-            .filter(|c| c.starts_with("/mod"))
-            .collect::<Vec<_>>();
-        assert!(
-            candidates.contains(&r.new_buffer.as_str()),
-            "{:?} must be one of the /mod* commands {:?} — no Frankenstein concatenation",
-            r.new_buffer,
-            candidates,
-        );
-        assert_eq!(
-            r.new_cursor,
-            r.new_buffer.len(),
-            "cursor should land at end of replacement",
-        );
-    }
-
-    /// Cursor at byte 0 (Home before Tab) used to produce
-    /// `/allow/mod` because the entire buffer was concatenated as
-    /// the tail. Verify it now produces a clean replacement —
-    /// exactly one of the matching commands, no residual `/mod`.
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn complete_with_cursor_at_start_produces_clean_buffer() {
-        let r = try_complete("/mod", 0).unwrap();
-        let candidates = builtin_commands()
-            .into_iter()
-            .filter(|c| c.starts_with("/mod"))
-            .collect::<Vec<_>>();
-        assert!(
-            candidates.contains(&r.new_buffer.as_str()),
-            "{:?} must be a /mod* command (clean replacement, no /mod residual): candidates {:?}",
-            r.new_buffer,
-            candidates,
-        );
-    }
-
-    /// Tab on a command with trailing args (e.g. `/mode standard`
-    /// with cursor after `/mode`) preserves the args tail.
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn complete_preserves_trailing_args() {
-        let r = try_complete("/mod standard", 4).unwrap();
-        assert!(
-            r.new_buffer.ends_with(" standard"),
-            "args after the command should be preserved: {:?}",
-            r.new_buffer
-        );
-    }
-
-    /// Cursor past the first whitespace means the user is typing
-    /// args, not a command name — no completion should fire.
-    #[cfg(feature = "experimental-ui-tab-slash")]
-    #[test]
-    fn no_completion_when_cursor_in_args() {
-        // Cursor inside the args portion of "/mode standard".
-        let buf = "/mode standard";
-        let cursor = buf.len(); // past the space
-        assert!(try_complete(buf, cursor).is_none());
-    }
-
-    // ============================================================
-    // Code-review B2 fix: canonical command list + drift guard
-    // ============================================================
-
-    /// `is_known_slash_command` must agree with `slash_command_names`
-    /// since the helper just iterates the list. Catches a future
-    /// refactor that decouples them (e.g. someone introducing a
-    /// second hardcoded match).
-    #[test]
-    fn is_known_slash_command_agrees_with_canonical_list() {
-        for name in slash_command_names() {
-            assert!(
-                is_known_slash_command(name),
-                "{name} is in slash_command_names() but is_known_slash_command rejects it",
-            );
-        }
-        // Spot-check negatives.
-        assert!(!is_known_slash_command("/not-a-real-command"));
-        assert!(!is_known_slash_command(""));
-        assert!(!is_known_slash_command("/"));
-    }
-
-    /// The canonical list is sorted (tab completion preview relies
-    /// on stable ordering for the cycle direction).
-    #[test]
-    fn slash_command_names_is_sorted() {
-        let cmds = slash_command_names();
-        for pair in cmds.windows(2) {
-            assert!(
-                pair[0] <= pair[1],
-                "{} should sort before {}",
-                pair[0],
-                pair[1]
-            );
-        }
-    }
-
-    /// Pin that the canonical list and `handle_slash`'s actual
-    /// match arms agree on the always-on commands. If a name in
-    /// the list above is missing from the dispatch tree the user
-    /// would hit the new "internal error" arm at runtime; this
-    /// duplicates the check in plain test code so a future
-    /// maintainer sees the gap before users do.
-    ///
-    /// This DOES NOT enforce the reverse direction (arm present in
-    /// `handle_slash` but missing from `slash_command_names`) — that
-    /// would require parsing source. We accept it as the lesser
-    /// drift: the only user-visible cost is that the missing
-    /// command isn't tab-completable.
-    #[test]
-    fn always_on_commands_appear_in_canonical_list() {
-        // Subset that is unconditionally compiled (no cfg) and
-        // therefore must always be present. Cross-checked by hand
-        // against the `match parts[0]` arms in `handle_slash`.
-        const ALWAYS_ON: &[&str] = &[
-            "/allow",
-            "/btw",
-            "/cd",
-            "/clear",
-            "/clone",
-            "/compact",
-            "/compress",
-            "/fork",
-            "/help",
-            "/mode",
-            "/model",
-            "/panel",
-            "/prompt",
-            "/quit",
-            "/reasoning",
-            "/regen-prompts",
-            "/retry",
-            "/sessions",
-            "/tasks",
-            "/toggle",
-            "/tree",
-            "/undo",
-        ];
-        let list = slash_command_names();
-        for name in ALWAYS_ON {
-            assert!(
-                list.contains(name),
-                "{name} must appear in slash_command_names() — it's an always-on dispatch arm in handle_slash",
-            );
-        }
-    }
-}
